Feat: 티켓 상태관리를 위한 useTicket 커스텀 훅 추가
