+++ src/App.tsx
@@ -2,36 +2,10 @@ import { useState } from "react";
 import { Ticket, TicketActions } from "./types";
 import TicketList from "./TicketList";
 import TicketForm from "./TicketForm";
+import { useTickets } from "./hooks/useTickets";
 
 function App() {
-  const [tickets, setTickets] = useState<Ticket[]>([]);
-
-  const ticketActions: TicketActions = {
-    addTicket({ title, description }) {
-      setTickets((prev) => [
-        ...prev,
-        {
-          id: prev.length > 0 ? Math.max(...prev.map((t) => t.id)) + 1 : 1,
-          title,
-          description,
-          status: "open",
-        },
-      ]);
-    },
-
-    toggleTicketStatus(id) {
-      setTickets((prev) =>
-        prev.map((ticket) =>
-          ticket.id === id
-            ? {
-                ...ticket,
-                status: ticket.status === "open" ? "closed" : "open",
-              }
-            : ticket
-        )
-      );
-    },
-  };
+  const { tickets, ticketActions } = useTickets();
 
   return (
     <div className="container">
